Split out test_post_url
